Bug fix: Allocator convex + better compile performance.

+++ src/draw/draw.cpp
@@ -46,10 +46,10 @@ void ekg::draw::rect(float x, float y, float w, float h, const ekg::vec4 &color,
     allocator.coord2f(1, 0);
     allocator.coord2f(0, 0);
 
-    data.material_color[0] = static_cast<int32_t>(color.x);
-    data.material_color[1] = static_cast<int32_t>(color.y);
-    data.material_color[2] = static_cast<int32_t>(color.z);
-    data.material_color[3] = static_cast<int32_t>(color.w);
+    data.material_color[0] = static_cast<uint8_t>(color.x);
+    data.material_color[1] = static_cast<uint8_t>(color.y);
+    data.material_color[2] = static_cast<uint8_t>(color.z);
+    data.material_color[3] = static_cast<uint8_t>(color.w);
     data.line_thickness = line_thickness;
     data.factor = 1;
